scripts/requirements.txt
dnacentersdk==2.0.2
webexteamssdk==1.6
attrdict==2.0.1
Jinja2==2.11.2
PyYAML==5.3.1
